Rename CommissionerControl to CommissionerControlDelegate
